+++ helm/oratio/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.3]
+### Added
+- Clock ability configuration
+
 ### Modified
 - Closed the oratio service to be accessible only from inside the cluster (i.e. by the kong service).